requirements.txt
beautifulsoup4==4.9.1
cached-property==1.5.1
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
idna==2.9
imageio==2.8.0
kiwisolver==1.2.0
lxml==4.5.1
matplotlib==3.2.2
networkx==2.4
numpy==1.19.0
opencv-python==4.3.0.36
path==14.0.1
Pillow==7.1.2
pyaml==20.4.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.24.0
robobrowser==0.5.3
scikit-image==0.17.2
scipy==1.5.0
six==1.15.0
soupsieve==2.0.1
tifffile==2020.6.3
urllib3==1.25.9
Werkzeug==1.0.1
